@@ -7,19 +7,19 @@
     "webextension-polyfill-ts": "0.26.0"
   },
   "devDependencies": {
-    "@babel/core": "7.19.3",
-    "@types/jasmine": "4.3.0",
-    "@types/jest": "29.1.1",
-    "babel-jest": "29.1.2",
-    "babel-loader": "8.2.5",
+    "@babel/core": "7.20.12",
+    "@types/jasmine": "4.3.1",
+    "@types/jest": "29.4.0",
+    "babel-jest": "29.4.3",
+    "babel-loader": "9.1.2",
     "file-loader": "6.2.0",
     "find": "0.3.0",
-    "jest": "29.1.2",
-    "ts-jest": "29.0.3",
-    "ts-loader": "9.4.1",
-    "typescript": "4.8.4",
-    "webpack": "5.74.0",
-    "webpack-cli": "4.10.0"
+    "jest": "29.4.3",
+    "ts-jest": "29.0.5",
+    "ts-loader": "9.4.2",
+    "typescript": "4.9.5",
+    "webpack": "5.75.0",
+    "webpack-cli": "5.0.1"
   },
   "jest": {
     "preset": "ts-jest"
