@@ -4,19 +4,19 @@
   "private": true,
   "homepage": "/options/",
   "dependencies": {
-    "@material-ui/core": "4.12.4",
-    "@types/jest": "29.1.1",
-    "@types/material-ui": "0.21.12",
-    "@types/node": "18.7.23",
-    "@types/react": "17.0.2",
-    "@types/react-dom": "18.0.6",
+    "@mui/material": "5.11.9",
+    "@emotion/styled": "11.10.6",
+    "@types/jest": "29.4.0",
+    "@types/node": "18.14.0",
+    "@types/react": "18.0.28",
+    "@types/react-dom": "18.0.11",
     "@types/react-router-dom": "5.3.3",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-axe": "3.5.4",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "6.4.1",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "6.8.1",
     "react-scripts": "5.0.1",
-    "typescript": "4.8.4",
+    "typescript": "4.9.5",
     "webextension-polyfill-ts": "0.22.0"
   },
   "scripts": {
